Finished restructuring information elements

+++ src/devices/mesh/802.11s/ie-dot11s-configuration.h
@@ -102,10 +102,9 @@ protected:
   {
     return IE11S_MESH_CONFIGURATION;
   }
-
-  uint16_t  GetInformationSize () const;
+  uint8_t  GetInformationSize () const;
   void SerializeInformation (Buffer::Iterator i) const;
-  uint16_t DeserializeInformation (Buffer::Iterator i, uint8_t length);
+  uint8_t DeserializeInformation (Buffer::Iterator i, uint8_t length);
   void PrintInformation(std::ostream& os) const;
   // TODO: Release and fill other fields in configuration
   // element
