+++ src/devices/mesh/802.11s/ie-dot11s-perr.h
@@ -59,9 +59,9 @@ protected:
     return IE11S_PERR;
   };
   void  SerializeInformation(Buffer::Iterator i) const;
-  uint16_t  DeserializeInformation(Buffer::Iterator start, uint8_t length);
+  uint8_t  DeserializeInformation(Buffer::Iterator start, uint8_t length);
   void PrintInformation(std::ostream& os) const;
-  uint16_t  GetInformationSize() const;
+  uint8_t  GetInformationSize() const;
 private:
   uint8_t   m_numOfDest;
   std::vector<HwmpRtable::FailedDestination>
