+++ src/devices/mesh/802.11s/ie-dot11s-prep.h
@@ -70,8 +70,8 @@ protected:
     return IE11S_PREP;
   }
   void SerializeInformation(Buffer::Iterator i) const;
-  uint16_t DeserializeInformation(Buffer::Iterator start, uint8_t length);
-  uint16_t GetInformationSize() const;
+  uint8_t DeserializeInformation(Buffer::Iterator start, uint8_t length);
+  uint8_t GetInformationSize() const;
   void PrintInformation(std::ostream& os) const;
 private:
   uint8_t  m_flags;
