+++ src/devices/mesh/802.11s/ie-dot11s-preq.h
@@ -108,8 +108,8 @@ protected:
     return IE11S_PREQ;
   }
   void SerializeInformation(Buffer::Iterator i) const;
-  uint16_t DeserializeInformation(Buffer::Iterator i, uint8_t length);
-  uint16_t GetInformationSize() const;
+  uint8_t DeserializeInformation(Buffer::Iterator i, uint8_t length);
+  uint8_t GetInformationSize() const;
   void PrintInformation(std::ostream& os) const;
 private:
   //how many destinations we support
